+++ services/Dtos/GetCategoriesDto.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace services.Dtos
+namespace RecipesAPI.Dtos
 {
     public class GetCategoriesDto
     {
