+++ services/Models/Base/BaseEntity.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace services.Models.Base
+namespace RecipesAPI.Models.Base
 {
     public class BaseEntity
     {
