+++ services/Models/RecipeIngredient.cs
@@ -1,10 +1,10 @@
-﻿using services.Models.Enums;
+﻿using RecipesAPI.Models.Enums;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace services.Models
+namespace RecipesAPI.Models
 {
     public class RecipeIngredient
     {
