+++ services/Program.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace services
+namespace RecipesAPI
 {
     public class Program
     {
