+++ services/RecipesAPI.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31729.503
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "services", "services.csproj", "{8872A859-68F7-4F67-B3CB-1A9455257270}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RecipesAPI", "RecipesAPI.csproj", "{8872A859-68F7-4F67-B3CB-1A9455257270}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
