+++ services/Startup.cs
@@ -8,14 +8,14 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
-using services.Data;
-using services.Service;
+using RecipesAPI.Data;
+using RecipesAPI.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace services
+namespace RecipesAPI
 {
     public class Startup
     {
